Rename _deolhonafila to deolhonafila, this is public now
